// src/film/manifest.gen.ts
export const FILM_GEN = {
  "version": 1,
  "count": 150,
  "padding": 3,
  "pattern": "f_%03d.webp",
  "ext": "webp",
  "crop": {
    "left": 0,
    "top": 0,
    "width": 2406,
    "height": 1354
  },
  "tiers": [
    {
      "id": "xl",
      "dir": "/film/1920",
      "width": 1920,
      "height": 1080,
      "totalBytes": 9439562,
      "avgBytes": 62930
    },
    {
      "id": "lg",
      "dir": "/film/1280",
      "width": 1280,
      "height": 720,
      "totalBytes": 5985198,
      "avgBytes": 39901
    },
    {
      "id": "sm",
      "dir": "/film/854",
      "width": 854,
      "height": 480,
      "totalBytes": 3751862,
      "avgBytes": 25012
    }
  ],
  "sourceMap": [
    1,
    3,
    5,
    7,
    9,
    11,
    13,
    15,
    17,
    19,
    21,
    23,
    25,
    27,
    29,
    31,
    33,
    35,
    37,
    39,
    41,
    43,
    45,
    47,
    49,
    51,
    53,
    55,
    57,
    59,
    61,
    63,
    65,
    67,
    69,
    71,
    73,
    75,
    77,
    79,
    81,
    83,
    85,
    87,
    89,
    91,
    93,
    95,
    97,
    99,
    101,
    103,
    105,
    107,
    109,
    111,
    113,
    115,
    117,
    119,
    121,
    123,
    125,
    127,
    129,
    131,
    133,
    135,
    137,
    139,
    141,
    143,
    145,
    147,
    149,
    151,
    153,
    155,
    157,
    159,
    161,
    163,
    165,
    167,
    169,
    171,
    173,
    175,
    177,
    179,
    181,
    183,
    185,
    187,
    189,
    191,
    193,
    195,
    197,
    199,
    201,
    203,
    205,
    207,
    209,
    211,
    213,
    215,
    217,
    219,
    221,
    223,
    225,
    227,
    229,
    231,
    233,
    235,
    237,
    239,
    241,
    243,
    245,
    247,
    249,
    251,
    253,
    255,
    257,
    259,
    261,
    263,
    265,
    267,
    269,
    271,
    273,
    275,
    277,
    279,
    281,
    283,
    285,
    287,
    289,
    291,
    293,
    295,
    297,
    299
  ],
  "heroFrames": [
    7,
    20,
    30,
    46,
    62,
    75,
    97,
    113,
    140
  ],
  "lqip": [
    "data:image/webp;base64,UklGRqQAAABXRUJQVlA4IJgAAAAQBACdASoYAA4APu1iqU2ppaOiMAgBMB2JQBdgMZPb+y90/yNwC4OcgAD+dZcRihKKf50mKwVDt3VFmWU6J4ZvcveN+IJ96gDzjysAK49F3cWZotmr9SnbXFEYtTO5DZMLrklDdlP/Ym7/hP8lD6jY2Rp468+FCdN+9pY/jkgZ84ekeCZDTmOOH76n8hlvC89euSs5MsCAAA==",
    "data:image/webp;base64,UklGRqQAAABXRUJQVlA4IJgAAADQAwCdASoYAA4APu1iqU2ppaOiMAgBMB2JQBOgA2ucWsqQCR/KaYgA8nv19p9IT8G0fFtag9PcBzVCpkg21ag1WubY4v2GcNY5FyED6hZg6gfzDWhsKUWLeFpNi8zTM02EJVm8Hv5ZI+TaYZyb73VCI5GxTEO1HgX2SO1xA20eV9bRPgkgpI5gKYd9Js5+4Zq00sMdaAAAAA==",
    "data:image/webp;base64,UklGRpgAAABXRUJQVlA4IIwAAABQBACdASoYAA4APu1iqU2ppaOiMAgBMB2JYwC/OdwAqXUXJrKd1BT9qy4AAP2blC5Q0b+VKjmRtPXzbWxLHEQ3Nn9Ct3Ip9GCINMVyAV7x+LcDyANNfXSFi/BF9GIWf2fwX+V3Zh6UpcdgvtyXgDRNBLM3RKCPjg2mW9JC2v+Ncum/0GIN64KBtdAAAA==",
    "data:image/webp;base64,UklGRqAAAABXRUJQVlA4IJQAAAAQBACdASoYAA4APu1iqU2ppaOiMAgBMB2JZQC06Yv02MD1d5SIg/l+AAD+zETMMKE7ZwMbCohaXv1FPeyMq6QQCx+q87bajv3o9Ut+4ovvjoXrvpovXjrHn6c2WFmQcHSISXSlEBtyWK+F9fAbAlKPMHNdum63IxQzRbPAQeSLUXhkBc/OuyN3CUg27UJ70dOVYAAA",
    "data:image/webp;base64,UklGRqYAAABXRUJQVlA4IJoAAADQAwCdASoYAA4APu1iqU2ppaOiMAgBMB2JQAIoACqa4W63YP7fF74A/r/5I8D5roh6gyBzTEOsihM7x80CnSSgkuIbpR9tFrrzr0VsU3vouMlcgRUJGiR8pfpnc39N38IG4gLV/1N//wRKAnl+FYGxpn6spNzYyz8wLu7IVBrWPRbapMxtJ8xH8q0fpLUHKqBucvEu5DizgAAA",
    "data:image/webp;base64,UklGRpIAAABXRUJQVlA4IIYAAABwAwCdASoYAA4APu1iqU2ppaOiMAgBMB2JYwCdABuzqQp1ZXAA/qnK7bPZOUoe7Y2TZ8pvxYXfL4JFpbntRnM4wdIaUdHTcjBgT4CPOLTcAK1eVDynOdnmodqQR1rrJgq9KGtCIzZ281SZUpXND+e3TAWQTDCm4S4p9wIiobzlaUPKqAAAAA==",
    "data:image/webp;base64,UklGRroAAABXRUJQVlA4IK4AAACQBACdASoYAA4APu1iqU2ppaOiMAgBMB2JZACdMoAlrBbh0Ys0mE+SQCWMusAA8p9XSLlJn0W4Xz8+Bd13S36bokHP7FPIjtLfyGFc0i66heDJXPvD9zl4/a7S/YkBz+FuArrSVVWIgw43IM9pMaQrSOeK1bQ9Qf4ypRx/J0sitFPaE4Kej8twPthCz+h3jpvO7bQ/73D1PguaIBY9v5np3fPbkCR5O0/UqbIAAAA=",
    "data:image/webp;base64,UklGRooAAABXRUJQVlA4IH4AAACwAwCdASoYAA4APu1iqU2ppaOiMAgBMB2JQBOgAoYh6alY6pLTEADiXIBS+335TcYB7IGL9UnxK6HCadtx9r9KhcikL720kUxfU0l1XjuUBm1Z3D9p8t8nFwyMt40ed/1SIo4/5zN9B9fXAFq1MA55p2yYAcxNYWwRfw1gAAA=",
    "data:image/webp;base64,UklGRqQAAABXRUJQVlA4IJgAAACQBACdASoYAA4APu1iqU2ppaOiMAgBMB2JYwCdIP8gwxGEx3wzCciTlk03JHAA/tEXac/+g2xUKlD0FmE4evXBTsoqM5dHlsMWjtNwNALr1NC4KD0ailQIwyIz1RO0p04jnRUxQcDD8+7qdBNBmvKGGrdmP1hcYkUCOqq1frMTbqJ6VDJzpavsbRvHyLy/z2TXwuLUMNIAAA=="
  ]
} as const